requirements.txt
folium
geopy
google-generativeai
plotly
sqlalchemy
streamlit
streamlit-lottie
streamlit_modal
# asyncore
# pycopy-asyncore
pyautogui